VB.Net
Example3 Rendition as I continue to go through these seems a little backwards to me, all the c++ examples the main loop is in the run procedure, from there your telling it to either update or draw, this interface all the ticks are coming from Draw. anyways just a quirk I noticed.